df_hgnn_project/requirements.txt

numpy
scipy
torch>=2.0
scikit-learn
pandas
pyyaml
